Initial restructuring for test suite
